x/trade/types/message_create_trade_test.go
package types

import (
	"testing"

	"github.com/GGEZLabs/testchain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateTrade_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateTrade
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateTrade{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateTrade{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
